Bit of code cleanup
